add ep16

+++ core/api_views.py
@@ -3,6 +3,8 @@ from .serializers import QuestionSerializer, AnswerSerializer
 from rest_framework.response import Response
 from rest_framework import status
 from .models import Question, Answer
+from rest_framework.permissions import IsAuthenticated
+from permissions import IsOwnerOrReadOnly
 
 
 class QuestionListView(APIView):
@@ -13,6 +15,8 @@ class QuestionListView(APIView):
 
 
 class QuestionCreateView(APIView):
+	permission_classes = [IsAuthenticated,]
+
 	def post(self, request):
 		data = QuestionSerializer(data=request.data)
 		if data.is_valid():
@@ -22,8 +26,11 @@ class QuestionCreateView(APIView):
 
 
 class QuestionUpdateView(APIView):
+	permission_classes = [IsOwnerOrReadOnly,]
+
 	def put(self, request, pk):
 		question = Question.objects.get(pk=pk)
+		self.check_object_permissions(request, question)
 		srz_data = QuestionSerializer(instance=question, data=request.data, partial=True)
 		if srz_data.is_valid():
 			srz_data.save()
@@ -32,7 +39,10 @@ class QuestionUpdateView(APIView):
 
 
 class QuestionDeleteView(APIView):
+	permission_classes = [IsOwnerOrReadOnly,]
+
 	def delete(self, request, pk):
 		question = Question.objects.get(pk=pk)
+		self.check_object_permissions(request, question)
 		question.delete()
 		return Response(status=status.HTTP_204_NO_CONTENT)
